dom/html/test/test_bug391777.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=391777
-->
<head>
  <title>Test for Bug 391777</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=391777">Mozilla Bug 391777</a>
<p id="display"></p>
<script class="testbody" type="text/javascript">

/** Test for Bug 391777 **/
SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv({ set: [[ "dom.disable_window_showModalDialog", false ]] }, doTest);
async function doTest() {
  let iframe = document.createElement("iframe");
  document.body.appendChild(iframe);
  await new Promise(resolve => {
    iframe.addEventListener("load", resolve);
  });
  window.showModalDialog = iframe.contentWindow.showModalDialog;
  var arg = {};
  arg.testVal = "foo";
  var result = window.showModalDialog("javascript:window.returnValue = window.dialogArguments.testVal; window.close(); 'This window should close on its own.';", arg);
  ok(true, "We should get here without user interaction");
  is(result, "foo", "Unexpected result from showModalDialog");
  SimpleTest.finish();
}

</script>
</body>
</html>
